backend/requirements.txt
# Backend dependencies for Render deployment
fastapi==0.115.0
uvicorn[standard]==0.30.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
Pillow==10.4.0
pytorch-grad-cam==1.5.4
ollama==0.4.0
python-multipart==0.0.9
